requirements.txt
yfinance
PyYAML
PyMySQL